+++ contributor-plugins/release/release-plugin/src/main/kotlin/com/mreil/easy/release/DefaultEasyReleaseExtension.kt
@@ -8,5 +8,6 @@ abstract class DefaultEasyReleaseExtension : EasyReleaseExtension {
     init {
         enabled.convention(true)
         releaseBranchPattern.convention("(main|master|rel-.*)")
+        preReleaseCommitMessage.convention("Set version for release: \$v")
     }
 }