+++ contributor-plugins/release/release-plugin/src/main/kotlin/com/mreil/easy/release/EasyReleasePlugin.kt
@@ -6,6 +6,7 @@ import com.mreil.easy.findEasyChild
 import com.mreil.easy.isRoot
 import com.mreil.easy.semver.EasySemver
 import com.mreil.easy.vcs.EasyVcs
+import com.mreil.easy.vcs.VcsType
 import com.mreil.utils.hasGroup
 import com.mreil.utils.hasVersion
 import org.gradle.api.Project
@@ -35,6 +36,13 @@ abstract class EasyReleasePlugin : AbstractEasyProjectPlugin() {
                 it.upToDate.set(vcs.flatMap { service -> service.isUpToDateWithRemote() })
                 it.commitSha.set(vcs.flatMap { service -> service.currentSha() })
             }
+        target.tasks.register("preReleaseCommit", PreReleaseCommitTask::class.java) {
+            it.group = "release"
+            it.description = "Writes the release version into the version file and commits it."
+            it.versionFile.set(release.versionFile.orElse(target.layout.projectDirectory.file("gradle.properties")))
+            it.commitMessageTemplate.set(release.preReleaseCommitMessage)
+            it.vcsType.set(vcs.map { service -> service.type() }.orElse(VcsType.NONE))
+        }
         target.tasks.register("release", Task::class.java) {
             it.group = "release"
             it.description = "Release wiring task (other release tasks attach here)."