+++ contributor-plugins/vcs/vcs-plugin-api/src/main/kotlin/com/mreil/easy/vcs/VcsNone.kt
@@ -19,4 +19,9 @@ internal class VcsNone(
     override fun push(): Provider<Boolean> = providers.provider { false }
 
     override fun fetch(): Provider<Boolean> = providers.provider { false }
+
+    override fun addAndCommit(
+        paths: List<String>,
+        message: String,
+    ): Provider<Boolean> = providers.provider { false }
 }